+++ lib/onlyoffice_documentserver_testing_framework/test_instance_docs/doc_editor.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'doc_editor/doc_editor_top_toolbar'
 require_relative 'doc_editor/doc_editor_windows'
 
 module OnlyofficeDocumentserverTestingFramework
@@ -7,6 +8,11 @@ module OnlyofficeDocumentserverTestingFramework
   class DocEditor
     def initialize(instance)
       @instance = instance
+    end
+
+    # @return [DocEditorTopToolbar] top toolbar
+    def top_toolbar
+      @top_toolbar ||= DocEditorTopToolbar.new(@instance)
     end
 
     def windows