+++ lib/onlyoffice_documentserver_testing_framework/test_instance_docs/spreadsheet_editor.rb
@@ -4,7 +4,7 @@ require_relative 'spreadsheet_editor/spreadsheet_windows'
 
 module OnlyofficeDocumentserverTestingFramework
   # Class for all Spreadsheet Editor actions
-  class SpreadsheetEditor
+  class SpreadsheetEditor < DocEditor
     def initialize(instance)
       @instance = instance
     end