+++ spec/presentation_editor/top_toolbar/top_toolbar_users/top_toolbar_users_spec.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe 'Management' do
+  let(:instance) { OnlyofficeDocumentserverTestingFramework::TestInstanceDocs.new }
+
+  before do
+    instance.webdriver.open('http://localhost')
+    instance.doc_service_welcome.go_to_example
+    instance.doc_test_functions.open_sample_document(:presentation)
+    instance.management.wait_for_operation_with_round_status_canvas
+  end
+
+  it 'default user button is not present' do
+    expect(instance.presentation_editor.top_toolbar.users).not_to be_present
+  end
+
+  it 'default user count is 1' do
+    expect(instance.presentation_editor.top_toolbar.users.count).to eq(1)
+  end
+
+  after do
+    instance.webdriver.quit
+  end
+end